+++ tezt/lib_tezos/baker.mli
@@ -126,9 +126,10 @@ val liquidity_baking_votefile : ?path:string -> liquidity_baking_vote -> string
     is not passed. If it is [Some x] then [--liquidity-baking-toggle-vote x] is
     passed. The default value is [Some Pass].
 
-    [operations_pool], [force_apply], [state_recorder],
+    [operations_pool], [force_apply_from_round], [state_recorder],
     [node_version_check_bypass] and [node_version_allowed] are passed to the
-    baker daemon through the flag [--operations-pool], [--force_apply],
+    baker daemon through the flag [--operations-pool],
+    [--force_apply_from_round],
     [--record-state], [--node-version-check-bypass] and
     [--node-version-allowed].
 
@@ -151,7 +152,7 @@ val create :
   ?delegates:string list ->
   ?votefile:string ->
   ?liquidity_baking_toggle_vote:liquidity_baking_vote option ->
-  ?force_apply:bool ->
+  ?force_apply_from_round:int ->
   ?remote_mode:bool ->
   ?operations_pool:string ->
   ?dal_node:Dal_node.t ->
@@ -192,7 +193,7 @@ val create_from_uris :
   ?delegates:string list ->
   ?votefile:string ->
   ?liquidity_baking_toggle_vote:liquidity_baking_vote option ->
-  ?force_apply:bool ->
+  ?force_apply_from_round:int ->
   ?remote_mode:bool ->
   ?operations_pool:string ->
   ?dal_node_rpc_endpoint:Endpoint.t ->
@@ -237,7 +238,7 @@ val create_from_uris :
     baker. This defaults to the empty list, which is a shortcut for "every known
     account".
 
-    [votefile], [liquidity_baking_toggle_vote], [force_apply],
+    [votefile], [liquidity_baking_toggle_vote], [force_apply_from_round],
     [state_recorder], [node_version_check_bypass], [node_version_allowed]
     respectively [operations_pool] are passed to the baker daemon through the
     flags [--votefile], [--liquidity-baking-toggle-vote], [--should-apply],
@@ -261,7 +262,7 @@ val init :
   ?delegates:string list ->
   ?votefile:string ->
   ?liquidity_baking_toggle_vote:liquidity_baking_vote option ->
-  ?force_apply:bool ->
+  ?force_apply_from_round:int ->
   ?remote_mode:bool ->
   ?operations_pool:string ->
   ?dal_node:Dal_node.t ->